refactor:  Simplified gender and birth year statements.
The Gender and Birth Year code was a complex statement to check for
errors when running calculations.  Replaced the code with a simple
conditional statement to check if gender and birth year columns
existed in the data before running calculations.

Changes suggested in the code review when I submitted the project.